backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
youtube-search==2.1.2
together==0.2.9
pydantic==2.6.1